Integrated basic chat functionality
- Currently navigable from the speech to text button, but that will be removed in the future
