+++ lib/services/bases/chat_bot.dart
@@ -2,10 +2,10 @@ import '../../models/ApplicationFunction.dart';
 import '../../utility/RouteGenerator.dart';
 
 /**
- * Abstract implementation for Chatbot Service functions
+ * Abstract implementation for ChatBot Service functions
  * to be publicly accessible
  */
-abstract class Chatbot {
+abstract class ChatBot {
   /**
    * Perform user entered chat function
    */
