--- lib/ui/chat.dart
@@ -1,194 +1,0 @@
-import 'dart:convert';
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart' show rootBundle;
-import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
-import 'package:flutter_chat_ui/flutter_chat_ui.dart';
-import 'package:http/http.dart' as http;
-import 'package:intl/date_symbol_data_local.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:uuid/uuid.dart';
-
-class ChatPage extends StatefulWidget {
-  const ChatPage({super.key});
-
-  @override
-  State<ChatPage> createState() => _ChatPageState();
-}
-
-class _ChatPageState extends State<ChatPage> {
-  List<types.Message> _messages = [];
-  final _user = const types.User(id: '82091008-a484-4a89-ae75-a22bf8d6f3ac');
-
-  FontWeight commonFontWt = FontWeight.w700;
-  double commonFontSize = 30;
-  double commonBorderWidth = 2;
-  double commonButtonHeight = 60;
-  double commonButtonWidth = 200;
-  double commonCornerRadius = 8;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadMessages();
-  }
-
-  @override
-  Widget build(BuildContext context) => Scaffold(
-    floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-    floatingActionButton: FloatingActionButton(
-      child: const Icon(Icons.mic),
-      onPressed: () {
-        // TODO
-      },
-    ),
-    bottomNavigationBar: BottomAppBar(
-      shape: CircularNotchedRectangle(),
-      notchMargin: 4.0,
-      child: Row(
-        mainAxisSize: MainAxisSize.max,
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: <Widget>[
-          Spacer(),
-          IconButton(
-            icon: Icon(Icons.headset_mic_outlined),
-            onPressed: () {},
-          )
-        ],
-      ),
-    ),
-    appBar: AppBar(
-      leading: GestureDetector(
-          onTap: () {
-            Navigator.pushNamed(context, '/main');
-          },
-          child: Icon(
-              Icons.arrow_back
-          )
-      ),
-      centerTitle: true,
-      title: Text(
-        "Chat Support",
-        style:
-        TextStyle(fontWeight: commonFontWt, fontSize: commonFontSize),
-      ),
-      actions: <Widget>[
-        Padding(
-            padding: EdgeInsets.only(right:20.0),
-            child: GestureDetector(
-                onTap: () {},
-                child: Icon(
-                    Icons.settings
-                )
-            )
-        ),
-        Padding(
-            padding: EdgeInsets.only(right:20.0),
-            child: GestureDetector(
-                onTap: () {},
-                child: Icon(
-                    Icons.logout
-                )
-            )
-        )
-      ],
-      automaticallyImplyLeading: false,
-      backgroundColor: Color(0xff004B87),
-    ),
-    body: Chat(
-      messages: _messages,
-      onMessageTap: _handleMessageTap,
-      onPreviewDataFetched: _handlePreviewDataFetched,
-      onSendPressed: _handleSendPressed,
-      showUserAvatars: true,
-      showUserNames: true,
-      user: _user,
-    ),
-  );
-
-  void _addMessage(types.Message message) {
-    setState(() {
-      _messages.insert(0, message);
-    });
-  }
-
-  void _handleMessageTap(BuildContext _, types.Message message) async {
-    if (message is types.FileMessage) {
-      var localPath = message.uri;
-
-      if (message.uri.startsWith('http')) {
-        try {
-          final index =
-          _messages.indexWhere((element) => element.id == message.id);
-          final updatedMessage =
-          (_messages[index] as types.FileMessage).copyWith(
-            isLoading: true,
-          );
-
-          setState(() {
-            _messages[index] = updatedMessage;
-          });
-
-          final client = http.Client();
-          final request = await client.get(Uri.parse(message.uri));
-          final bytes = request.bodyBytes;
-          final documentsDir = (await getApplicationDocumentsDirectory()).path;
-          localPath = '$documentsDir/${message.name}';
-
-          if (!File(localPath).existsSync()) {
-            final file = File(localPath);
-            await file.writeAsBytes(bytes);
-          }
-        } finally {
-          final index =
-          _messages.indexWhere((element) => element.id == message.id);
-          final updatedMessage =
-          (_messages[index] as types.FileMessage).copyWith(
-            isLoading: null,
-          );
-
-          setState(() {
-            _messages[index] = updatedMessage;
-          });
-        }
-      }
-    }
-  }
-
-  void _handlePreviewDataFetched(
-      types.TextMessage message,
-      types.PreviewData previewData,
-      ) {
-    final index = _messages.indexWhere((element) => element.id == message.id);
-    final updatedMessage = (_messages[index] as types.TextMessage).copyWith(
-      previewData: previewData,
-    );
-
-    setState(() {
-      _messages[index] = updatedMessage;
-    });
-  }
-
-  void _handleSendPressed(types.PartialText message) {
-    final textMessage = types.TextMessage(
-      author: _user,
-      createdAt: DateTime.now().millisecondsSinceEpoch,
-      id: const Uuid().v4(),
-      text: message.text,
-    );
-
-    _addMessage(textMessage);
-  }
-
-  void _loadMessages() async {
-    final response = await rootBundle.loadString('assets/messages.json');
-    final messages = (jsonDecode(response) as List)
-        .map((e) => types.Message.fromJson(e as Map<String, dynamic>))
-        .toList();
-
-    setState(() {
-      _messages = messages;
-    });
-  }
-}
