+++ lib/utility/RouteGenerator.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:summer2022/ui/chat_widget.dart';
 import 'package:summer2022/ui/mail_widget.dart';
 import 'package:summer2022/ui/main_menu.dart';
 import 'package:summer2022/ui/other_mail.dart';
@@ -12,7 +13,10 @@ import 'package:summer2022/models/EmailArguments.dart';
 enum SiteAreas { Home, Settings, Search, SearchResults, MailView, NotificationManage, NotificationView }
 
 class RouteGenerator {
+  static SiteAreas previousRoute = SiteAreas.Home;
+
   static Route<dynamic> generateRoute(RouteSettings settings) {
+    _updatePreviousRoute(settings.name!);
     print('setting.name:  ${settings.name}');
     switch (settings.name) {
       case '/main':
@@ -32,6 +36,11 @@ class RouteGenerator {
       case '/backend_testing':
         return MaterialPageRoute(
             builder: (_) => const BackendPage(title: "Backend Testing"));
+      case '/chat':
+        return MaterialPageRoute(
+            builder: (_) => ChatWidget(
+              currentPage: previousRoute
+        ));
       default:
         return MaterialPageRoute(builder: (_) {
           return Scaffold(
@@ -39,6 +48,24 @@ class RouteGenerator {
             child: Text("No route for ${settings.name}"),
           ));
         });
+    }
+  }
+
+  /**
+   * Update previous route to new route if it matches any valid routes
+   * This is for use by the ChatBot to determine the user's page
+   * TODO: Probably a better way to handle this with Navigator/navkey
+   */
+  static void _updatePreviousRoute(String newRoute) {
+    switch (newRoute) {
+      case '/main':
+        previousRoute = SiteAreas.Home;
+        break;
+      case '/settings':
+        previousRoute = SiteAreas.Settings;
+        break;
+      default:
+        break;
     }
   }
 }
