+++ pubspec.yaml
@@ -65,7 +65,6 @@ dev_dependencies:
     sdk: flutter
   build_runner: ^2.1.11
   flutter_launcher_icons: "^0.9.2"
-  mockito: ^5.3.1
 
 flutter_icons:
  android: true
